Move HexInput to /components
