+++ src/js/utils/hextorgb.js
@@ -1,7 +1,6 @@
 var hexToRgb = {
 
     convert: function (hex) {
-        console.log("Converting: " + hex);
 
         var r = (parseInt(hex.substring(0,2), 16)/255).toFixed(2);
         var g = (parseInt(hex.substring(2,4), 16)/255).toFixed(2);
